+++ src/main/kotlin/com/github/igr/sexyeditor/SexyEditor.kt
@@ -1,3 +1,0 @@
-package com.github.igr.sexyeditor
-
-const val COMPONENT_NAME = "SexyEditor"