+++ src/main/kotlin/com/github/igr/sexyeditor/config/PluginState.kt
@@ -1,6 +1,5 @@
 package com.github.igr.sexyeditor.config
 
-import com.github.igr.sexyeditor.COMPONENT_NAME
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
@@ -8,30 +7,30 @@ import com.intellij.openapi.components.Storage
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 @State(
-    name = COMPONENT_NAME,
-    storages = [Storage("sexy-editor-3.xml")]
+    name = "AwesomeEditor",
+    storages = [Storage("awesome-editor-3.xml")]
 )
-class SexyEditorState : PersistentStateComponent<SexyEditorState> {
+class PluginState : PersistentStateComponent<PluginState> {
 
     var configs: MutableList<BackgroundConfiguration> = mutableListOf()
 
     companion object {
-        fun get(): SexyEditorState {
-            return ApplicationManager.getApplication().getService(SexyEditorState::class.java)
+        fun get(): PluginState {
+            return ApplicationManager.getApplication().getService(PluginState::class.java)
         }
     }
 
     /**
      * Returns plugin state.
      */
-    override fun getState(): SexyEditorState {
+    override fun getState(): PluginState {
         return this
     }
 
     /**
      * Loads state from configuration file.
      */
-    override fun loadState(state: SexyEditorState) {
+    override fun loadState(state: PluginState) {
         XmlSerializerUtil.copyBean(state, this)
     }
 }